Switch times to long and radii to doubles

+++ src/main/java/edu/brown/cs/drawbridge/carpools/Carpools.java
@@ -72,8 +72,8 @@ public final class Carpools {
    * @return A List of valid paths. Each path is a List of Trips.
    */
   public List<List<Trip>> searchWithId(String userId, double startLat,
-      double startLon, double endLat, double endLon, int departureTime,
-      int walkingTime, int timeRadius) {
+      double startLon, double endLat, double endLon, long departureTime,
+      double walkingTime, double timeRadius) {
     return tripSearcher
         .searchWithId(userId, startLat, startLon, endLat, endLon, departureTime,
             walkingTime * WALKING_SPEED, timeRadius * SECONDS_PER_MINUTE);
@@ -103,8 +103,8 @@ public final class Carpools {
    * @return A List of valid paths. Each path is a List of Trips.
    */
   public List<List<Trip>> searchWithoutId(double startLat, double startLon,
-      double endLat, double endLon, int departureTime, int walkingTime,
-      int timeRadius) {
+      double endLat, double endLon, long departureTime, double walkingTime,
+      double timeRadius) {
     return tripSearcher
         .searchWithoutId(startLat, startLon, endLat, endLon, departureTime,
             walkingTime * WALKING_SPEED, timeRadius * SECONDS_PER_MINUTE);
